Move RKEClusterSpecCommon to cluster_configuration_types.go
